improv: code cleanup

+++ src/main/java/com/aetherteam/aether/item/combat/abilities/armor/PhoenixArmor.java
@@ -21,6 +21,8 @@ import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.phys.Vec3;
 import net.neoforged.neoforge.event.entity.living.LivingEvent;
+import net.neoforged.neoforge.event.entity.living.LivingIncomingDamageEvent;
+import net.neoforged.neoforge.event.tick.EntityTickEvent;
 
 public interface PhoenixArmor {
     /**
@@ -28,7 +30,7 @@ public interface PhoenixArmor {
      * Wearing Phoenix Armor also clears any fire from the wearer and spawns flame particles around them.
      *
      * @param entity The {@link LivingEntity} wearing the armor.
-     * @see com.aetherteam.aether.event.listeners.abilities.ArmorAbilityListener#onEntityUpdate(LivingEvent.LivingTickEvent)
+     * @see com.aetherteam.aether.event.listeners.abilities.ArmorAbilityListener#onEntityUpdate(EntityTickEvent.Post)
      */
     static void boostLavaSwimming(LivingEntity entity) {
         if (EquipmentUtil.hasFullPhoenixSet(entity)) {
@@ -39,7 +41,7 @@ public interface PhoenixArmor {
                     var data = player.getData(AetherDataAttachments.AETHER_PLAYER);
                     float defaultBoost = boostWithDepthStrider(entity, 1.75F, 1.0F);
                     data.setPhoenixSubmergeLength(Math.min(data.getPhoenixSubmergeLength() + 0.1, 1.0));
-                    defaultBoost *= data.getPhoenixSubmergeLength();
+                    defaultBoost *= (float) data.getPhoenixSubmergeLength();
                     entity.moveRelative(0.04F * defaultBoost, new Vec3(entity.xxa, entity.yya, entity.zza));
                 } else {
                     float defaultBoost = boostWithDepthStrider(entity, 1.75F, 1.0F);
@@ -65,7 +67,7 @@ public interface PhoenixArmor {
      * Boosts the entity's vertical movement in lava if wearing a full set of Phoenix Armor. The default boost is modified based on duration in lava and whether the boots have Depth Strider.<br><br>
      *
      * @param entity The {@link LivingEntity} wearing the armor.
-     * @see com.aetherteam.aether.event.listeners.abilities.ArmorAbilityListener#onEntityUpdate(LivingEvent.LivingTickEvent)
+     * @see com.aetherteam.aether.event.listeners.abilities.ArmorAbilityListener#onEntityUpdate(EntityTickEvent.Post)
      */
     static void boostVerticalLavaSwimming(LivingEntity entity) {
         if (EquipmentUtil.hasFullPhoenixSet(entity)) {
@@ -76,7 +78,7 @@ public interface PhoenixArmor {
                     var data = player.getData(AetherDataAttachments.AETHER_PLAYER);
                     float defaultBoost = boostWithDepthStrider(entity, 1.5F, 0.05F);
                     data.setPhoenixSubmergeLength(Math.min(data.getPhoenixSubmergeLength() + 0.1, 1.0));
-                    defaultBoost *= data.getPhoenixSubmergeLength();
+                    defaultBoost *= (float) data.getPhoenixSubmergeLength();
                     if (entity.getDeltaMovement().y() > 0 || entity.isCrouching()) {
                         entity.setDeltaMovement(entity.getDeltaMovement().multiply(1.0, defaultBoost, 1.0));
                     }
@@ -113,7 +115,7 @@ public interface PhoenixArmor {
      * The methods used for this are {@link PhoenixArmor#breakPhoenixArmor(LivingEntity, ItemStack, ItemStack, EquipmentSlot)} and {@link PhoenixArmor#breakPhoenixGloves(LivingEntity, SlotEntryReference, ItemStack)}.
      *
      * @param entity The {@link LivingEntity} wearing the armor.
-     * @see com.aetherteam.aether.event.listeners.abilities.ArmorAbilityListener#onEntityUpdate(LivingEvent.LivingTickEvent)
+     * @see com.aetherteam.aether.event.listeners.abilities.ArmorAbilityListener#onEntityUpdate(EntityTickEvent.Post)
      */
     static void damageArmor(LivingEntity entity) {
         if (entity instanceof Player player) {
@@ -192,7 +194,7 @@ public interface PhoenixArmor {
      * @param entity The {@link LivingEntity} wearing the armor.
      * @param source The attacking {@link DamageSource}.
      * @return Whether the fire damage should be cancelled, as a {@link Boolean}.
-     * @see com.aetherteam.aether.event.listeners.abilities.ArmorAbilityListener#onEntityAttack(LivingAttackEvent)
+     * @see com.aetherteam.aether.event.listeners.abilities.ArmorAbilityListener#onEntityAttack(LivingIncomingDamageEvent)
      */
     static boolean extinguishUser(LivingEntity entity, DamageSource source) {
         return EquipmentUtil.hasFullPhoenixSet(entity) && source.is(DamageTypeTags.IS_FIRE);
